+++ test/units/prawn/emoji/image_test.rb
@@ -9,14 +9,29 @@ describe Prawn::Emoji::Image do
 
     it 'possible to find the image file' do
       emoji_unicodes.each do |emoji_unicode|
-        emoji_image = Prawn::Emoji::Image.new(emoji_unicode)
+        emoji_image = Prawn::Emoji::Image.new(emoji_unicode, 12)
         assert File.exist?(emoji_image.path), "#{emoji_unicode}  not found"
       end
     end
   end
 
+  describe '#adjust_x' do
+    subject { Prawn::Emoji::Image.new('🍣', 12).adjust_x(100) }
+    it { subject.must_equal 100.9 }
+  end
+
+  describe '#adjust_y' do
+    subject { Prawn::Emoji::Image.new('❤', 18).adjust_y(100) }
+    it { subject.must_equal 114.4 }
+  end
+
+  describe '#width' do
+    subject { Prawn::Emoji::Image.new('❤', 10).width }
+    it { subject.must_equal 8.5 }
+  end
+
   describe '#==' do
-    it { Prawn::Emoji::Image.new('🍣').must_be :==, Prawn::Emoji::Image.new('🍣') }
-    it { Prawn::Emoji::Image.new('🍣').wont_be :==, Prawn::Emoji::Image.new('🐟') }
+    it { Prawn::Emoji::Image.new('🍣', 12).must_be :==, Prawn::Emoji::Image.new('🍣', 12) }
+    it { Prawn::Emoji::Image.new('🍣', 12).wont_be :==, Prawn::Emoji::Image.new('🐟', 12) }
   end
 end