+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Breaking Changes:
  * prawn-emoji requires Prawn 2.2 (Drop support Prawn 2.1)
  * Improve emoji support for ZWJ/Combining Sequence and Text Presentation
    * See https://github.com/hidakatsuya/prawn-emoji/blob/master/test/pdf/emoji_rendering/expect.pdf
+ * Switch emoji image to [Twemoji](https://github.com/twitter/twemoji) v12.04 from EmojiOne v2.0
 
 ## 2.1.1